Initial Commit
Orbits 3.9 - Final product of a turtle based 3D n-body simulator.
Orbits 4T - Also a turtle based simulator, but uses more advanced
mathematics to allow for projection of objects onto the screen with a
camera position and rotation independent of the objects position,
anywhere in space.

Orbits 4 - Hopefully will be the same as 4T but not turtle based, making
it faster.