Potentially working product
I found a pretty big flaw in the maths I used to calculate the X and Y
position. After redoing it, I got better equations and now then
particles show potentially correct movement on screen. The major and
minor axes calculations haven't been redone, I've just gone back to the
original working ones. I'll have to do some more testing to know if this
is working or not.

Also, I've updated some of Projection.pdf to show the updated maths.
Note that the rewrite is not done, and somewhere there is probably a
sudden change in topic.