Visuals are looking good
The graphical part of the program is looking promising. Started
implementing panning and rotation, the first of which has turned out to
be more difficult than I thought.
Also, for some reason I've made a merge earlier on. Github was acting up
and being really annoying and I'm not really sure what dangerous things
I've done while messing around with it to fix it.