Added z-axis rotation, moved vector class into separate file
I've added the ability for rotation about the screen's z-axis, which
turned out to be easier than I'd expected. Also added a few keyboard
functions, like stepping delta and screen depth up and down. The program
was getting pretty big so I've moved the vector class, a hefty 220 ish
lines, into a separate file. vector.py is already another library I've
made and is in a different repo, but this one is more updated and can be
considered the latest version.