Solar System Replication!
Preset 3 Now loads in the solar system, with data from a SolSystem.txt.
Currently I have data for all the planets but not for the Moon or Pluto.
The code to load the system, while small, has been put in a separate
program (I might start moving more functions/classes etc. to other
prorams, orbits4T.py is getting pretty lengthy). Only problems with this
so far is that the camera panning is always either aggressively fast or
depressingly slow. Since the solar system is quite big, I'll need to
make more changes to the autoRate algorithm to better determine a pan
rate. Also, turns out things in the solar system are hard to see unless
you're literally right next to them. I've added an option 'draw stars'
(enabled with '-ds') which draws white dots if the normal shape is too
small to see.