Improvements to display data
The data shown on the screen now shows distances and speeds with
relevant prefixes, reducing the length of the number making it more
readable. Also added a 'relative speed to camera' figure, which taught
me that there isn't really an easy way to find the velocity of the
camera at the moment. That could do with some attention.
Also, I added the New Horizons spacecraft and the ISS.