LeapFrog included, Added test mode
Attempted to implement leapfrog integration into the calculation of
speed and position. Not sure if it's working correctly.

Also added a feature for testing the simulation by comparing computed
positions and velocitities with values currently hard coded into the
program (I might put them in a separate file later on). So far it shows
that any bodies significantly close to another, or with a highly
changing acceleration, generally have the largest difference between the
simulated and correct outcome. For example, the ISS is often thousands
of kilometres away from it's correct location after 1 orbit if the delta
is much larger than 1 second.