Added Stars!
the command line argument -getStars followed by nothing or an integer
will now load 20 or up to just under 120,000 stars from the night sky!
Of course, I wouldn't ask that much of any mere mortal computer, but
it's cool to think that its there. The stars currently just sit
motionless, using the new 'static' property of the particle. But the
data I got includes the velocities, so if I want I can make them move.
Also, I've added parsec as a prefix in the prefixing function, but I
think I broke it. So fix that.
Also, I have started giving the camera an actual velocity, with mixed
results.