Load System is more complicated
The loadFile function of loadSystem now allows the text file to state
the type of delimiter used. In the StarsData.txt file for example, the
lines ~DELIM and ~REQUIRED both allow the user to specify the delimiter
the program should be looking for and the columns that are required to
make an entry 'valid'.