Star Brightness!
The brightness of stars has been added, with apparent magnitude being
correctly calculated based on distance. This looks great when you're
really far away from the star but the script to draw up the flare is
pretty basic, and doesn't look very good as you approach an object. But
it does allow, for example, constellations such as the Southern Cross to
shine out over the other stars.
Also, there is a pretty serious glitch which literally locks up the
whole computer which I am almost certain comes from this feature. When
you are close enough to a star to see it and it's flare, and then rotate
away from the star, when the star is (i assume) exactly perpendicular to
your heading/rotation then it freezes up, and sometimes the next frame
will be all white/yellow/red depending on the colour of the star. I
think it's probably trying to draw a ridiculously large flare.