Added reflections for planets etc. and slightly better polydraw
Planets and pretty much anthing designated as a reflector can now 'reflect' light incident on them an basically appear to have an apparent magnitude. Kinda works but it's really weird.

Also the oval drawing function is now finally able to clip bits of the oval that aren't on screen, hopefully helping speed up the program a bit.